220907 pruebas2

+++ utils.c
@@ -6,7 +6,7 @@
 /*   By: bjimenez <bjimenez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/31 09:32:31 by bjimenez          #+#    #+#             */
-/*   Updated: 2022/09/06 23:01:38 by bjimenez         ###   ########.fr       */
+/*   Updated: 2022/09/07 15:40:36 by bjimenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,9 +46,9 @@ long int	ft_timenow(void)
 	return ((now.tv_sec * 1000) + (now.tv_usec / 1000));
 }
 
-void	ft_delay(long int delay)
+void	ft_delay(float delay)
 {
-	long int	time;
+	float	time;
 
 	time = ft_timenow() + delay;
 	while (time >= ft_timenow())
